requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
blinker==1.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
honeycomb-beeline==2.17.0
idna==2.10
importlib-metadata==2.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
libhoney==1.10.0
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pyee==8.1.0
pyppeteer==0.2.5
python-dotenv==0.18.0
regex==2021.4.4
requests==2.25.1
sentry-sdk==1.1.0
six==1.16.0
statsd==3.3.0
toml==0.10.2
tqdm==4.61.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
uWSGI==2.0.19.1
websockets==8.1
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
